// apps/web/src/lib/api/generated/models/resultAnalysis.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Formex
 * OpenAPI spec version: 0.0.1
 */
import type { Student } from "./student";
import type { Subject } from "./subject";

export interface ResultAnalysis {
  id: string;
  course_id: string;
  student_id: string;
  scored_grade: string;
  created_at: string;
  updated_at: string;
  course?: Subject;
  student?: Student;
}